rename example files
