+++ src/library_src/hello.cpp
@@ -1,2 +1,2 @@
-#include "lol.hpp"
+#include <hello.hpp>
 void fun() {int i = 0; ++i;}
