+++ tests_src/hello_test/helloTest.cpp
@@ -1,12 +1,12 @@
 #define BOOST_TEST_NO_LIB
 #include <boost/test/unit_test.hpp>
 
-#include "lol.hpp"
+#include "hello.hpp"
 
 BOOST_AUTO_TEST_CASE(first_test)
 {
 fun();
-  int i = 1;
+  int i = 14;
   BOOST_TEST(i);
   BOOST_TEST(i == 2);
 }
